requirements.txt
requests==2.31.0
cryptography==41.0.7
pyyaml==6.0.1
pydantic==2.5.0
python-socketio[client]==5.9.0
click==8.1.7
colorama==0.4.6
psutil==5.9.6
watchdog==3.0.0
python-magic==0.4.27
python-magic-bin==0.4.14
chardet==5.2.0
tqdm==4.66.1
jsonschema==4.20.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Build tools
pyinstaller==5.13.2